New feature: sorting of .proto files (#20)
The order in which .proto files are given to the compiler is file system dependent, and thus by extension also operating system dependent, which doesn't matter much
for code generation, but can be critical for documentation generation (for which protoc plugins exist!). This commit adds an optional parameter "sortProtoFiles" which can be set to true (defaults to false) to have the .proto files be sorted lexicographically according to their path and file names.